% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/climbingStrategy.R
\docType{methods}
\name{climbingStrategy}
\alias{climbingStrategy}
\title{The climbingStrategy method}
\usage{
climbingStrategy(object)
}
\arguments{
\item{object:}{an object of HillClimbing class}
}
\value{
landscape.max: the maximum profit for the local landscape
}
\description{
climbingStrategy() is a method created for HillClimbing
class which uses three different strategies SA,MA and LA to find the
maximum profit on a local economic landscape
}

