% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/findNeighborhood.R
\docType{methods}
\name{findNeighborhood}
\alias{findNeighborhood}
\title{The findNeighborhood method}
\usage{
findNeighborhood(object)
}
\arguments{
\item{object:}{an object of HillClimbing class}
}
\value{
neighbors: a matrix which contains all neighbors of the initial allocation
}
\description{
findNeighborhood() is a method created for HillClimbing
class which finds all neighbors of a given allocation in the landscape
}

