% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generateObject.R
\docType{methods}
\name{generateObject}
\alias{generateObject}
\title{The generateObject method}
\usage{
generateObject(object)
}
\arguments{
\item{object:}{an object of HillClimbing class}
}
\value{
object: an object of HillClimbing class
}
\description{
generateObject() is a method created for HillClimbing class which gives
the object the random starting allocation and randomized coefficients of the profit function
}

