% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HillClimbing.R
\docType{class}
\name{HillClimbing-class}
\alias{HillClimbing-class}
\title{The HillClimbing class}
\description{
The HillClimbing class is a class which contains slots used by
different functions in the package
}
\section{Slots}{

\describe{
\item{\code{approach:}}{three simple strategies for hillclimbing firms, "SA", "MA" or "LA"}

\item{\code{budget:}}{the amount of firm's total budget}

\item{\code{input:}}{the number of total inputs}

\item{\code{connection:}}{the number of connections per input}

\item{\code{allocation:}}{a vector which contains a set of inputs subject to the budget constraint}

\item{\code{linearConstraint:}}{a vector which gives constraints for linear coefficients}

\item{\code{squareConstraint:}}{a vector which gives constraints for square coefficients}

\item{\code{crossproductConstraint:}}{a vector which gives constraints for cross product coefficients}

\item{\code{coef1:}}{a vector which stores all linear coefficients of the profit function}

\item{\code{coef2:}}{a matrix which stores all quadratic coefficients
(both square and cross product coefficients) of the profit function}
}}
\author{
Fan Zhang
}

