% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/neighborhoodProfit.R
\docType{methods}
\name{neighborhoodProfit}
\alias{neighborhoodProfit}
\title{The neighborhoodProfit method}
\usage{
neighborhoodProfit(object)
}
\arguments{
\item{object:}{an object of HillClimbing class}
}
\value{
neighbors.profit: a vector which contains normalized profits for all neighbors
}
\description{
neighborhoodProfit() is a method created for
"HillClimbing" class which calculates normalized profits for all neighbors
}

