% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/paperResult.R
\docType{methods}
\name{paperResult}
\alias{paperResult}
\title{The paperResult method}
\usage{
paperResult(object)
}
\arguments{
\item{object:}{an object of HillClimbing class}
}
\value{
result: the mean normalized profit over 1000 economic landscapes
}
\description{
paperResult() is a method created for HillClimbing
class which reproduces results from the three tables in the paper
}

