% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/profitFunction.R
\docType{methods}
\name{profitFunction}
\alias{profitFunction}
\title{The profitFunction method}
\usage{
profitFunction(object)
}
\arguments{
\item{object:}{an object of "HillClimbing" class}
}
\value{
profit.norm: the normalized profit for the allocation
}
\description{
profitFunction() is a method created for HillClimbing
class which calculates the normalized profit for any given allocation.
}

